Added a 2nd follicle to the straight path
Fixed a major bug! Added a 2nd follicle on the straight path which then stabilizes the sliding vector when animating sideways. If animating sideways across the path, the controls would drift. This was never caught in testing because the test was animated straight in Z.

Also unpacked oClosePoint into the 3 components the function returns.